산출물/V. 파인튜닝/requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
bitsandbytes>=0.39.0
scipy
numpy
tqdm
peft>=0.4.0
sentencepiece
python-dotenv
huggingface_hub 